Conflicts:
	Sublime Text 3/Preferences.sublime-settings

+++ install.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 # link dotfiles to home directory
 
+# get the OS type
+OS=$(uname -s)
+
 # get current working directory
 CWD=$(cd "$(dirname "$0")"; pwd)
 
@@ -12,12 +15,17 @@ ln -s $CWD/.zshrc ~/.zshrc
 rm ~/.vimrc
 ln -s $CWD/.vimrc ~/.vimrc
 
-# For Sublime Text 3
-_IFS_=$IFS
-IFS="|"
-SUBLIME="$HOME/Library/Application Support/Sublime Text 3/Packages/User"
-rm $SUBLIME/Python.sublime-build
-ln -s "$CWD/Sublime Text 3/Python.sublime-build" $SUBLIME/Python.sublime-build
-rm $SUBLIME/Preferences.sublime-settings
-ln -s "$CWD/Sublime Text 3/Preferences.sublime-settings" $SUBLIME/Preferences.sublime-settings
-IFS=$_IFS_
+if [ $OS = "Darwin" ]; then
+	# For Sublime Text 3
+	_IFS_=$IFS
+	IFS="|"
+	SUBLIME="$HOME/Library/Application Support/Sublime Text 3/Packages/User"
+	rm $SUBLIME/Python.sublime-build
+	ln -s "$CWD/Sublime Text 3/Python.sublime-build" $SUBLIME/Python.sublime-build
+	rm $SUBLIME/Preferences.sublime-settings
+	ln -s "$CWD/Sublime Text 3/Preferences.sublime-settings" $SUBLIME/Preferences.sublime-settings
+	IFS=$_IFS_
+else
+	echo "Linux no need Sublime Text 3"
+fi
+
